third_party/blink/web_tests/platform/generic/fast/files/not-enough-arguments-expected.txt

This test tries calling various FileAPI functions with too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS reader.readAsArrayBuffer() threw exception TypeError: Failed to execute 'readAsArrayBuffer' on 'FileReader': 1 argument required, but only 0 present..
PASS reader.readAsBinaryString() threw exception TypeError: Failed to execute 'readAsBinaryString' on 'FileReader': 1 argument required, but only 0 present..
PASS reader.readAsText() threw exception TypeError: Failed to execute 'readAsText' on 'FileReader': 1 argument required, but only 0 present..
PASS reader.readAsDataURL() threw exception TypeError: Failed to execute 'readAsDataURL' on 'FileReader': 1 argument required, but only 0 present..
PASS fileInput.files.item() threw exception TypeError: Failed to execute 'item' on 'FileList': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
